update_crewai/requirements.txt
crewai>=0.28.8
langchain>=0.1.20
langchain-openai>=0.1.0
langchain-community>=0.0.38
pymongo>=4.6.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0
chromadb>=0.4.22
tiktoken>=0.6.0
pydantic>=2.6.0
numpy>=1.24.0
